@@ -22,8 +22,8 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.0-alpha14'
-        classpath 'com.google.gms:google-services:3.1.0'
+        classpath 'com.android.tools.build:gradle:3.3.1'
+        classpath 'com.google.gms:google-services:4.2.0'
     }
 }
 
@@ -35,16 +35,15 @@ allprojects {
 }
 
 ext {
-    buildTools = '27.0.3'
-    compileSdk = 27
+    compileSdk = 28
     minSdk = 15
     versionCode = 1
     versionName = '1.0'
-    supportLib = '27.0.1'
+    appCompat = '1.0.2'
     firebaseVer = '11.6.0'
     instantAppsVer = '1.1.0'
 
-    espressoVersion = "3.0.1"
-    androidTestVersion = "1.0.1"
+    espressoVersion = '3.1.0'
+    androidTestVersion = '1.1.0'
     hamcrestVersion = "1.3"
 }
