@@ -16,14 +16,14 @@
 
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = '1.1.51'
+    ext.kotlin_version = '1.3.10'
     repositories {
-        jcenter()
         google()
+        jcenter()
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.0-alpha14'
+        classpath 'com.android.tools.build:gradle:3.3.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
         // NOTE: Do not place your application dependencies here; they belong
@@ -33,20 +33,19 @@ buildscript {
 
 allprojects {
     repositories {
-        jcenter()
         google()
+        jcenter()
     }
 }
 
 ext {
-    buildTools = '27.0.3'
-    compileSdk = 27
+    compileSdk = 28
     minSdk = 15
     versionCode = 1
     versionName = "1.0"
-    supportLib = '27.0.1'
-
-    espressoVersion = "3.0.1"
-    androidTestVersion = "1.0.1"
+    appCompat = "1.0.2"
+    constraintlayout = "1.1.3"
+    espressoVersion = '3.1.0'
+    androidTestVersion = '1.1.0'
     hamcrestVersion = "1.3"
 }
