@@ -17,18 +17,18 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 ext {
-    buildTools = '27.0.3'
-    compileSdk = 27
-    targetSdk = 27
+    compileSdk = 28
+    targetSdk = 28
     minSdk = 21
     versionCode = 1
     versionName = "1.0"
-    supportLib = '27.0.1'
+    appCompat = '1.0.2'
+    recyclerView = '1.0.0'
     glide = "3.7.0"
     retrofit = "1.9.0"
 
-    espressoVersion = "3.0.1"
-    androidTestVersion = "1.0.1"
+    espressoVersion = '3.1.0'
+    androidTestVersion = '1.1.0'
     hamcrestVersion = "1.3"
 }
 
@@ -38,7 +38,7 @@ buildscript {
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.0-alpha14'
+        classpath 'com.android.tools.build:gradle:3.3.1'
     }
 }
 allprojects {
