@@ -22,7 +22,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.0-alpha14'
+        classpath 'com.android.tools.build:gradle:3.3.1'
     }
 }
 
@@ -34,15 +34,14 @@ allprojects {
 }
 
 ext {
-    buildTools = '27.0.3'
-    compileSdk = 27
+    compileSdk = 28
     minSdk = 23
     versionCode = 1
     versionName = '1.0'
-    supportLib = '27.0.1'
-    constraintLayout = '1.0.2'
+    appCompat = '1.0.2'
+    constraintLayout = '1.1.3'
 
-    espressoVersion = "3.0.1"
-    androidTestVersion = "1.0.1"
+    espressoVersion = '3.1.0'
+    androidTestVersion = '1.1.0'
     hamcrestVersion = "1.3"
 }
